deus_rest/requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.4
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lxml==4.3.4
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
pandas==0.24.2
pandas-datareader==0.7.0
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.7.6.1
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.0
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.2
virtualenv==16.1.0
virtualenv-clone==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
